Refactor getParamsAndProps code (#7537)
