+++ packages/astro/src/core/render/index.ts
@@ -1,11 +1,7 @@
 export { createRenderContext } from './context.js';
 export type { RenderContext } from './context.js';
-export {
-	getParamsAndProps,
-	GetParamsAndPropsError,
-	getParamsAndPropsOrThrow,
-	renderPage,
-} from './core.js';
+export { renderPage } from './core.js';
 export type { Environment } from './environment';
 export { createBasicEnvironment, createEnvironment } from './environment.js';
+export { getParamsAndProps } from './params-and-props.js';
 export { loadRenderer, loadRenderers } from './renderer.js';
